Create Bionest.md
Bionest is the Biohacker Space of the Innovation Park Agrobioteg. Innovation Park Agrobioteg is the Innovation HUB most important in Guanajuato, in this place we support innovation projects in Life Sciences. We are hostess of more of five Biotechnology Startups. Bionest works as a "seedbed" of projects where ideas are materialized.